Add files via upload
This technology is used for smart" Attendance recognition and detection in classroom" generally teachers note the attendance orally and mark on sheet, as technology is developed there must be a change in education system as well. Our product makes use of an camera component which recognizes the faces of the students and mark the attendance itself so, this makes the work faster and even simpler . In order to improve the advancement in attendance system we have introduced this technique which reduces the work of teachers in classroom.